Modified and Solved mainly Rooms, Dimension classes
