+++ TeachingAssistant.java
@@ -1,3 +1,4 @@
+
 import java.util.Arrays;
 
 public class TeachingAssistant extends Staff {
