+++ enums/Equipment.java
@@ -1,0 +1,8 @@
+package enums;
+public enum Equipment {
+    NONE,
+    SMART_BOARD,
+    WHITEBOARD,
+    COMPUTERS,
+    PROJECTOR
+}
